// src/types/heroTheme.ts
export interface HeroThemeConfig {
    darkSlide1Media?: string;
    darkSlide1Video?: string;
    darkSlide1Image?: string;

    darkSlide2Media?: string;
    darkSlide2Video?: string;
    darkSlide2Image?: string;

    darkPreset?: string; // 'royal_gold' | 'luminous_gold' | 'sovereign_silver' | 'cinematic'

    lightSlide1Media?: string;
    lightSlide1Video?: string;
    lightSlide1Image?: string;

    lightSlide2Media?: string;
    lightSlide2Video?: string;
    lightSlide2Image?: string;

    lightPreset?: string; // 'royal_gold' | 'luminous_gold' | 'sovereign_silver' | 'cinematic'
}